unify store tests and reduce code duplication
